Chapter 5 Fixes (4)
I see the thing with the 0s in the csv. That must be something that Excel automatically does? I'll try to keep those out of the commit. It doesn't seem to break anything, but a small change to the csv in labeling 0x8A as "Chapter 5 Generic." In my "Custom Definitions.event," my generic is 0x8A.