Chapter 5 Changes (#41)
* Chapter 5 Changes

Tilechanges still do not work, and the second half of the map is untested.

* Chapter 5 Fixes

I hope to add +10 crit to whoever is holding the Power Gem when the time for the second batch of chapters comes around since I won't have one.

* Add note of lyn in setup instructions.

* Chapter 5 Fixes (2)

* Chapter 5 Fixes (3)

* Chapter 5 Fixes (4)

I see the thing with the 0s in the csv. That must be something that Excel automatically does? I'll try to keep those out of the commit. It doesn't seem to break anything, but a small change to the csv in labeling 0x8A as "Chapter 5 Generic." In my "Custom Definitions.event," my generic is 0x8A.

* Chapter 5 Fixes (5)

Oof

* Remove PowerGem in production verions

* Update Chapter 5.txt

* Chaptr 5 Fixes (6)

* Chaptr 5 Changes (7)

* Fixed Ch5 tile issue.

* Fixed village destroyed issue.