character stuff
Added names and descriptions for generic enemies that I forgot.
Wicky's dark rank and Splenda's light rank are intentional; normal witch class has anima. This is to reflect the good/wicked witch thing.
Benny is supposed to be an Est-like unit.
Zane is well-rounded prepromote.
Splenda is fast and accurate, but not very strong. Still have to implement the witch skill.