Repoint portrait table and transplant vanilla table. (#115)
* Repoint portrait table and transplant vanilla table.

* Create Mug Status.txt

I made a list of all the characters who need mugs. I just did this in game, and a fair amount of portraits are just not assigned to the correct slots yet.
(I'll throw it in this branch because it's dedicated to portraits.

* Fixing overwritten portraits

Portraits already present in the buildfile yet not present in game for whatever reason have been fixed.

* Adding pre-claimed mugs

All mugs on the mug sheet that have been claimed by somebody have been added to the game. Also importred the few not-imported VBA1 mugs.

* Portrait repoint error fix (#117)

* Reverting newly added IDs

They're all 0x01 for testing purposes. The game still crashes when viewing class cards.

* Update VanillaPortraitTable.dmp

Copies 0xAC worth of portraits instead of 0x80. This seems to repair that crashing bug.

* Revert "Reverting newly added IDs"

This reverts commit bffd322439be253b0a32444ad2939febef903029.

* Merge palettes into portrait branch (#118)

* Update Battle Palettes.txt

added more palettes for player and boss units